RC-v1.7.1: Endow designation (#1688)
* endow designation param

* move logic to custom hook

* add designations filter

* add designation type

* slice upts

* add level2
