+++ src/slices/components/marketFilter/types.ts
@@ -1,4 +1,4 @@
-import { EndowmentsSortKey, SortDirection } from "types/aws";
+import { EndowDesignation, EndowmentsSortKey, SortDirection } from "types/aws";
 import { CapitalizedEndowmentType, EndowmentTier } from "types/contracts";
 import { UNSDG_NUMS } from "types/lists";
 
@@ -12,6 +12,7 @@ export type FilterState = {
   isOpen: boolean;
   searchText: string;
   endow_types: CapitalizedEndowmentType[];
+  endow_designation: EndowDesignation[];
   sort?: Sort;
   sdgGroups: SdgGroups;
   region: { activities: Regions; headquarters: Regions };
